% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourly_grid.R
\name{hourly_grid}
\alias{hourly_grid}
\title{Retrieve the Hourly Grid File for a Given Date and Time}
\usage{
hourly_grid(hour, time_format, dir)
}
\arguments{
\item{hour}{Character. Date and time used to identify the grid file.}

\item{time_format}{Character. Format used to parse \code{hour}.}

\item{dir}{Character. Directory containing the hourly grid shapefiles.}
}
\value{
A character string containing the name of the matching shapefile.
}
\description{
Searches a directory containing hourly pollutant grid shapefiles and
returns the filename corresponding to the requested date and hour.
}
\details{
Hourly grid shapefiles are expected to follow the naming convention
\code{"YYYY-mm-dd_HHMM.shp"}.
}
\examples{
\dontrun{
hourly_grid(
  hour = "2019-08-01 08:00:00",
  time_format = "\%Y-\%m-\%d \%H:\%M:\%S",
  dir = "path/to/grids"
)
}

}
