% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_grid.R
\name{make_grid}
\alias{make_grid}
\title{Create example hourly pollutant grids}
\usage{
make_grid(ymin, ymax, xmin, xmax, pixelSize, Dir, date, values = NULL)
}
\arguments{
\item{ymin}{Numeric. Minimum latitude of the grid extent.}

\item{ymax}{Numeric. Maximum latitude of the grid extent.}

\item{xmin}{Numeric. Minimum longitude of the grid extent.}

\item{xmax}{Numeric. Maximum longitude of the grid extent.}

\item{pixelSize}{Numeric. Grid cell size in meters.}

\item{Dir}{Character. Directory where the shapefiles will be written.}

\item{date}{Character. Date used in the output filenames
(\code{"YYYY-mm-dd"}).}

\item{values}{Optional. Currently unused.}
}
\value{
This function is called for its side effect of writing shapefiles to disk.
No value is returned.
}
\description{
Creates a regular spatial grid over a user-defined bounding box and exports
24 hourly shapefiles. Each grid cell is assigned a random value, making this
function useful for generating example datasets and testing workflows.
}
\details{
The function creates one shapefile for each hour of the day. Pollutant
values are generated randomly and are intended only for examples and
testing.
}
\examples{
\dontrun{
make_grid(
  ymin = -31.6,
  ymax = -31.3,
  xmin = -64.3,
  xmax = -64.0,
  pixelSize = 1000,
  Dir = "example_grids/",
  date = "2019-08-01"
)
}

}
